+++ src/SampleTest/Linq/EnumerableSequenceEqualTest.cs
@@ -34,4 +34,15 @@ public class EnumerableSequenceEqualTest {
 		// Assert
 		Assert.Equal(expected, actual);
 	}
+
+	[Fact]
+	public void SequenceEqual_2つのコレクションが等しいか比較できる() {
+		// Arrange
+		var first = new[] { 1, 2, 3 };
+		var second = new[] { 1, 2, 3 };
+
+		// Act
+		// Assert
+		Assert.True(first.SequenceEqual(second));
+	}
 }
