+++ src/SampleTest/Linq/EnumerableTest.cs
@@ -95,76 +95,4 @@ public class EnumerableTest {
 		// null許容型の場合はnull
 		Assert.Null(Enumerable.Empty<int?>().Average());
 	}
-
-	[Fact]
-	public void Take_引数にマイナスの値を指定すると空のシーケンスが返ってくる() {
-		// Arrange
-		var source = new[] { 1, 2, 3 };
-
-		// Act
-		var actual = source.Take(-1);
-
-		// Assert
-		Assert.Empty(actual);
-	}
-
-	[Fact]
-	public void Take_Range型で指定してシーケンスを取得する() {
-		// Arrange
-		var source = new[] { 0, 1, 2, 3, 4, 5 };
-
-		// Act
-		// 2番目から4番目まで（4番目は含まない）を取得
-		var actual = source.Take(2..4);
-
-		// Assert
-		Assert.Equal([2, 3], actual);
-	}
-
-	[Fact]
-	public void Take_Range型で指定したが要素が見つからない場合は空が返ってくる() {
-		// Arrange
-		var source = new[] { 0, 1, 2, 3 };
-
-		// Act
-		// 2番目から4番目まで（4番目は含まない）を取得
-		var actual = source.Take(5..);
-
-		// Assert
-		Assert.Empty(actual);
-	}
-
-	[Fact]
-	public void Skip_引数にマイナスの値を指定すると同じシーケンスが返ってくる() {
-		// Arrange
-		var source = new[] { 1, 2, 3 };
-
-		// Act
-		var actual = source.Skip(-1);
-
-		// Assert
-		Assert.NotSame(source, actual);
-		Assert.Equal(source, actual);
-	}
-
-	[Fact]
-	public void SequenceEqual_2つのコレクションが等しいか比較できる() {
-		// Arrange
-		var first = new[] { 1, 2, 3 };
-		var second = new[] { 1, 2, 3 };
-
-		// Act
-		// Assert
-		Assert.True(first.SequenceEqual(second));
-	}
-
-	[Fact]
-	public void ToList_クラスのListをインターフェイスのIListに変換する() {
-		// Arrange
-		var items = new List<Sample> { new() };
-
-		// Act
-		// Assert
-		Assert.True(items.ToList<ISample>() is List<ISample>);
-	}
 }
